se modifico las vistas para la carga de convenios
