+++ frontend/src/components/convenio/gestionconvenios.component.js
@@ -13,7 +13,7 @@ import Box from '@material-ui/core/Box';
 
 import Convenio from './convenioabm.component'
 import Categoria from './categorias.component'
-
+import SumasRemunerativas from './sumasremunerativas.component'
 
 function TabPanel(props) {
     const { children, value, index, ...other } = props;
@@ -107,6 +107,7 @@ export default function GestionarConvenios(props) {
                             <Tabs value={value} onChange={handleChange} aria-label="simple tabs example">
                                 <Tab label="Convenios" {...a11yProps(0)} />
                                 <Tab label="Categorias" {...a11yProps(1)} />
+                                <Tab label="Sumas remunerativas" {...a11yProps(2)} />
                                 
                             </Tabs>
                         </AppBar>
@@ -115,6 +116,9 @@ export default function GestionarConvenios(props) {
                         </TabPanel>
                         <TabPanel value={value} index={1} className={classes.tabs}>
                             <Categoria/>
+                        </TabPanel>
+                        <TabPanel value={value} index={2} className={classes.tabs}>
+                            <SumasRemunerativas/>
                         </TabPanel>
                         
                     </div>
